ScreenShot/requirements.txt
python-docx
pyautogui
pynput
Pillow
